♻️ Make minor improvements
See https://github.com/alrra/dotfiles/commit/4d23d0c3874597598f21e94ed2ae762b213f93f3

+++ src/os/restart.sh
@@ -1,5 +1,23 @@
 #!/usr/bin/env bash
 
-sudo shutdown -r now &> /dev/null
-#              |  └─ Immediately (alternative is +number, or yymmddhhmm)
-#              └──── The system is rebooted at the specified time.
+cd "$(dirname "${BASH_SOURCE[0]}")" \
+    && . "utils.sh"
+
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+main() {
+
+    print_in_purple "\n\n * Restart\n\n"
+
+    ask_for_confirmation "Do you want to restart?"
+    printf "\n"
+
+    if answer_is_yes; then
+        sudo shutdown -r now &> /dev/null
+        #              |  └─ Immediately (alternative is +number, or yymmddhhmm)
+        #              └──── The system is rebooted at the specified time.
+    fi
+
+ }
+
+ main
